Moved enums from Game to Game.Enums
