+++ Shelter/Game/CostumeConverter.cs
@@ -1,4 +1,5 @@
 using ExitGames.Client.Photon;
+using Game.Enums;
 using Mod;
 using UnityEngine;
