+++ Shelter/Game/Enums/BottomType.cs
@@ -1,4 +1,4 @@
-namespace Game
+namespace Game.Enums
 {
     public enum BottomType
     {
