+++ Shelter/Game/Enums/CheckPointState.cs
@@ -1,4 +1,4 @@
-namespace Game
+namespace Game.Enums
 {
     public enum CheckPointState
     {
