+++ Shelter/Game/Enums/Daylight.cs
@@ -1,4 +1,4 @@
-namespace Game
+namespace Game.Enums
 {
     public enum Daylight
     {
