+++ Shelter/Game/Enums/UniformType.cs
@@ -1,4 +1,4 @@
-namespace Game
+namespace Game.Enums
 {
     public enum UniformType
     {
