+++ Shelter/Game/GameManager.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using Game;
+using Game.Enums;
 using Mod.GameSettings;
 using Mod.Keybinds;
 using Mod.Managers;
