+++ Shelter/Game/HERO_DEAD_BODY_SETUP.cs
@@ -1,4 +1,5 @@
 using Game;
+using Game.Enums;
 using UnityEngine;
 
 // ReSharper disable once CheckNamespace
