+++ Shelter/Game/HeroCostume.cs
@@ -1,3 +1,4 @@
+using Game.Enums;
 using UnityEngine;
 
 namespace Game
