+++ Shelter/Game/Interface/BTN_choose_human.cs
@@ -1,6 +1,7 @@
 using System.Linq;
 using Game;
 using ExitGames.Client.Photon;
+using Game.Enums;
 using JetBrains.Annotations;
 using Mod;
 using NGUI.Internal;
