+++ Shelter/Game/Interface/BTN_choose_titan.cs
@@ -1,5 +1,6 @@
 using Game;
 using ExitGames.Client.Photon;
+using Game.Enums;
 using JetBrains.Annotations;
 using Mod;
 using NGUI.Internal;
