+++ Shelter/Game/LevelInfoManager.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using Game.Enums;
 using UnityEngine;
 
 namespace Game
