+++ Shelter/Game/PVPcheckPoint.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Game;
+using Game.Enums;
 using JetBrains.Annotations;
 using Mod;
 using Photon;
