+++ Shelter/Mod/PlayerProperties.cs
@@ -1,5 +1,6 @@
 ﻿using Game;
 using ExitGames.Client.Photon;
+using Game.Enums;
 using UnityEngine;
 
 namespace Mod
